refactor: split terminal presentation helpers
