+++ src/ui/features/terminal/Terminal.tsx
@@ -34,11 +34,10 @@ import { OPKSSHDialog } from "@/ssh/dialogs/OPKSSHDialog.tsx";
 import { HostKeyVerificationDialog } from "@/ssh/dialogs/HostKeyVerificationDialog.tsx";
 import { TmuxSessionPicker } from "@/ssh/dialogs/TmuxSessionPicker.tsx";
 import {
-  TERMINAL_THEMES,
   DEFAULT_TERMINAL_CONFIG,
   TERMINAL_FONTS,
 } from "@/lib/terminal-themes.ts";
-import type { TerminalConfig } from "@/types";
+import "./terminal-global-styles.ts";
 import { useTheme } from "@/components/theme-provider.tsx";
 import { useCommandTracker } from "@/features/terminal/command-history/useCommandTracker.ts";
 import { highlightTerminalOutput } from "@/lib/terminal-syntax-highlighter.ts";
@@ -53,72 +52,9 @@ import {
 import { ConnectionLog } from "@/ssh/connection-log/ConnectionLog.tsx";
 import { toast } from "sonner";
 import { Button } from "@/components/button";
-
-// Background/foreground per UI theme for "Termix Default" — must match index.css
-const TERMIX_DEFAULT_COLORS: Record<
-  string,
-  { background: string; foreground: string }
-> = {
-  dark: { background: "#0c0d0b", foreground: "#fafafa" },
-  light: { background: "#ffffff", foreground: "#111210" },
-  dracula: { background: "#282a36", foreground: "#f8f8f2" },
-  catppuccin: { background: "#1e1e2e", foreground: "#cdd6f4" },
-  nord: { background: "#2e3440", foreground: "#eceff4" },
-  solarized: { background: "#002b36", foreground: "#839496" },
-  "tokyo-night": { background: "#1a1b26", foreground: "#a9b1d6" },
-  "one-dark": { background: "#282c34", foreground: "#abb2bf" },
-  gruvbox: { background: "#282828", foreground: "#ebdbb2" },
-};
-
-function resolveTermixThemeColors(activeTheme: string, appTheme: string) {
-  if (activeTheme !== "termix") {
-    return (
-      TERMINAL_THEMES[activeTheme]?.colors || TERMINAL_THEMES.termixDark.colors
-    );
-  }
-  let resolvedUiTheme = appTheme;
-  if (appTheme === "system") {
-    resolvedUiTheme = window.matchMedia("(prefers-color-scheme: dark)").matches
-      ? "dark"
-      : "light";
-  }
-  const uiColors =
-    TERMIX_DEFAULT_COLORS[resolvedUiTheme] ?? TERMIX_DEFAULT_COLORS.dark;
-  const base = TERMINAL_THEMES.termixDark.colors;
-  return {
-    ...base,
-    background: uiColors.background,
-    foreground: uiColors.foreground,
-    cursor: uiColors.foreground,
-    cursorAccent: uiColors.background,
-  };
-}
-
-export interface TerminalHostConfig {
-  id?: number;
-  instanceId?: string;
-  restoredSessionId?: string | null;
-  ip: string;
-  port: number;
-  username: string;
-  password?: string;
-  key?: string;
-  keyPassword?: string;
-  keyType?: string;
-  authType?: string;
-  credentialId?: number;
-  terminalConfig?: TerminalConfig;
-  [key: string]: unknown;
-}
-
-export interface TerminalHandle {
-  disconnect: () => void;
-  reconnect: () => void;
-  fit: () => void;
-  sendInput: (data: string) => void;
-  notifyResize: () => void;
-  refresh: () => void;
-}
+import { resolveTermixThemeColors } from "./terminal-theme.ts";
+import type { TerminalHandle, TerminalHostConfig } from "./terminal-types.ts";
+export type { TerminalHandle, TerminalHostConfig } from "./terminal-types.ts";
 
 type HostKeyVerificationData = Omit<
   React.ComponentProps<typeof HostKeyVerificationDialog>,
@@ -2776,85 +2712,3 @@ export const Terminal = forwardRef<TerminalHandle, SSHTerminalProps>(
     );
   },
 );
-
-const style = document.createElement("style");
-style.innerHTML = `
-@import url('https://fonts.googleapis.com/css2?family=JetBrains+Mono:ital,wght@0,400;0,700;1,400;1,700&display=swap');
-@import url('https://fonts.googleapis.com/css2?family=Fira+Code:wght@400;700&display=swap');
-@import url('https://fonts.googleapis.com/css2?family=Source+Code+Pro:ital,wght@0,400;0,700;1,400;1,700&display=swap');
-
-@font-face {
-  font-family: 'Caskaydia Cove Nerd Font Mono';
-  src: url('./fonts/CaskaydiaCoveNerdFontMono-Regular.ttf') format('truetype');
-  font-weight: normal;
-  font-style: normal;
-  font-display: swap;
-}
-
-@font-face {
-  font-family: 'Caskaydia Cove Nerd Font Mono';
-  src: url('./fonts/CaskaydiaCoveNerdFontMono-Bold.ttf') format('truetype');
-  font-weight: bold;
-  font-style: normal;
-  font-display: swap;
-}
-
-@font-face {
-  font-family: 'Caskaydia Cove Nerd Font Mono';
-  src: url('./fonts/CaskaydiaCoveNerdFontMono-Italic.ttf') format('truetype');
-  font-weight: normal;
-  font-style: italic;
-  font-display: swap;
-}
-
-@font-face {
-  font-family: 'Caskaydia Cove Nerd Font Mono';
-  src: url('./fonts/CaskaydiaCoveNerdFontMono-BoldItalic.ttf') format('truetype');
-  font-weight: bold;
-  font-style: italic;
-  font-display: swap;
-}
-
-.xterm .xterm-viewport::-webkit-scrollbar {
-  width: 8px;
-  background: transparent;
-}
-.xterm .xterm-viewport::-webkit-scrollbar-thumb {
-  background: rgba(0,0,0,0.3);
-  border-radius: 4px;
-}
-.xterm .xterm-viewport::-webkit-scrollbar-thumb:hover {
-  background: rgba(0,0,0,0.5);
-}
-.xterm .xterm-viewport {
-  scrollbar-width: thin;
-  scrollbar-color: rgba(0,0,0,0.3) transparent;
-}
-
-.dark .xterm .xterm-viewport::-webkit-scrollbar-thumb {
-  background: rgba(255,255,255,0.3);
-}
-.dark .xterm .xterm-viewport::-webkit-scrollbar-thumb:hover {
-  background: rgba(255,255,255,0.5);
-}
-.dark .xterm .xterm-viewport {
-  scrollbar-color: rgba(255,255,255,0.3) transparent;
-}
-
-.xterm {
-  font-feature-settings: "liga" 0, "calt" 0;
-  text-rendering: optimizeLegibility;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-.xterm .xterm-screen {
-  font-family: 'Caskaydia Cove Nerd Font Mono', 'SF Mono', Monaco, Consolas, 'Liberation Mono', 'Courier New', monospace !important;
-  font-variant-ligatures: none;
-}
-
-.xterm .xterm-screen .xterm-char {
-  font-feature-settings: "liga" 0, "calt" 0;
-}
-`;
-document.head.appendChild(style);
