+++ src/ui/features/terminal/terminal-global-styles.ts
@@ -1,0 +1,81 @@
+const style = document.createElement("style");
+style.innerHTML = `
+@import url('https://fonts.googleapis.com/css2?family=JetBrains+Mono:ital,wght@0,400;0,700;1,400;1,700&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Fira+Code:wght@400;700&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Source+Code+Pro:ital,wght@0,400;0,700;1,400;1,700&display=swap');
+
+@font-face {
+  font-family: 'Caskaydia Cove Nerd Font Mono';
+  src: url('./fonts/CaskaydiaCoveNerdFontMono-Regular.ttf') format('truetype');
+  font-weight: normal;
+  font-style: normal;
+  font-display: swap;
+}
+
+@font-face {
+  font-family: 'Caskaydia Cove Nerd Font Mono';
+  src: url('./fonts/CaskaydiaCoveNerdFontMono-Bold.ttf') format('truetype');
+  font-weight: bold;
+  font-style: normal;
+  font-display: swap;
+}
+
+@font-face {
+  font-family: 'Caskaydia Cove Nerd Font Mono';
+  src: url('./fonts/CaskaydiaCoveNerdFontMono-Italic.ttf') format('truetype');
+  font-weight: normal;
+  font-style: italic;
+  font-display: swap;
+}
+
+@font-face {
+  font-family: 'Caskaydia Cove Nerd Font Mono';
+  src: url('./fonts/CaskaydiaCoveNerdFontMono-BoldItalic.ttf') format('truetype');
+  font-weight: bold;
+  font-style: italic;
+  font-display: swap;
+}
+
+.xterm .xterm-viewport::-webkit-scrollbar {
+  width: 8px;
+  background: transparent;
+}
+.xterm .xterm-viewport::-webkit-scrollbar-thumb {
+  background: rgba(0,0,0,0.3);
+  border-radius: 4px;
+}
+.xterm .xterm-viewport::-webkit-scrollbar-thumb:hover {
+  background: rgba(0,0,0,0.5);
+}
+.xterm .xterm-viewport {
+  scrollbar-width: thin;
+  scrollbar-color: rgba(0,0,0,0.3) transparent;
+}
+
+.dark .xterm .xterm-viewport::-webkit-scrollbar-thumb {
+  background: rgba(255,255,255,0.3);
+}
+.dark .xterm .xterm-viewport::-webkit-scrollbar-thumb:hover {
+  background: rgba(255,255,255,0.5);
+}
+.dark .xterm .xterm-viewport {
+  scrollbar-color: rgba(255,255,255,0.3) transparent;
+}
+
+.xterm {
+  font-feature-settings: "liga" 0, "calt" 0;
+  text-rendering: optimizeLegibility;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+.xterm .xterm-screen {
+  font-family: 'Caskaydia Cove Nerd Font Mono', 'SF Mono', Monaco, Consolas, 'Liberation Mono', 'Courier New', monospace !important;
+  font-variant-ligatures: none;
+}
+
+.xterm .xterm-screen .xterm-char {
+  font-feature-settings: "liga" 0, "calt" 0;
+}
+`;
+document.head.appendChild(style);
